Update Serializer_Deserializer.v
mask the upper 4 bits of each data sample, to remove any inconsistencies with reading the MSB. It's the 1st bit read from the ADC, and is always coming out of tri-state so may read as a 1 or 0. The ADC provides 12-bit data samples and pads the uppermost 4 bits with 0's.